+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog
 
+## Version 0.1.1
+
+- Support of Odoo 10.0
+- Support of [setuptools-odoo](https://pypi.python.org/pypi/setuptools-odoo)
+  - Automaticaly install in env
+  - Wrap pip with automaticaly set `PIP_EXTRA_INDEX_URL` environment variable with [OCA Wheelhouse](https://wheelhouse.odoo-community.org/)
+- Added shortcut script `odoo-helper-restart` to restart server.
+- Added `odoo-helper db rename` command
+- Added `odoo-helper install reinstall-venv` option
+- `odoo-helper test`: Test only installable addons
+- `odoo-helper addons update-list` command support odoo 7.0
+- `odoo-helper addons test-installed` command support odoo 7.0
+- `odoo-helper fetch`: added experimental support of Mercurial
+- `odoo-helper test --coverage-html` option added.
+- `odoo-helper db create` new options added:
+  - `--demo` load demo data (default: not load)
+  - `--lang <lang` choose language of database
+  - `--help` display help message
+- `odoo-install --single-branch` option added. This allow to disable `single-branch` clone.
+- Added `pychart` for install
+  `pychart` package is broken on pypi, so replace it with Python-Chart
+
+
 ## Version 0.1.0
 
 - Added ``odoo-helper addons pull_updates`` command